Static analysis: fix C6340 sign mismatch in nav date formatting
cl /analyze pass over the project translation units (TextEditor.cpp/.h,
editor.cpp, main.cpp, selftest.cpp) was clean except one finding:
SYSTEMTIME fields (WORD / unsigned short) passed to "%04d-%02d..." in
snprintf. Harmless at runtime (varargs promotes to int) but a real sign
mismatch — cast each to int explicitly. Re-analyze confirms editor.cpp now
clean; all four project TUs pass /analyze.

+++ example/editor.cpp
@@ -1220,7 +1220,8 @@ static void navRenderEntry(Editor* self,
 				FileTimeToSystemTime(&ft, &st);
 				char buf[64];
 				std::snprintf(buf, sizeof(buf), "%04d-%02d-%02d %02d:%02d",
-					st.wYear, st.wMonth, st.wDay, st.wHour, st.wMinute);
+					(int) st.wYear, (int) st.wMonth, (int) st.wDay,
+					(int) st.wHour, (int) st.wMinute);
 				return buf;
 			};
 			ImGui::TextDisabled("Created:  %s", fmt(fad.ftCreationTime).c_str());
